enhance: move the search untilities from querynode into new module
