+++ internal/util/searchutil/optimizers/mock_query_hook.go
