+++ internal/util/searchutil/optimizers/query_hook.go
