+++ internal/util/searchutil/optimizers/query_hook_test.go
