+++ internal/util/searchutil/scheduler/concurrent_safe_scheduler.go
@@ -1,4 +1,4 @@
-package tasks
+package scheduler
 
 import (
 	"fmt"
