+++ internal/util/searchutil/scheduler/mock_task_test.go
@@ -1,4 +1,4 @@
-package tasks
+package scheduler
 
 import (
 	"context"
