+++ internal/util/searchutil/scheduler/queues.go
@@ -1,4 +1,4 @@
-package tasks
+package scheduler
 
 import (
 	"container/ring"
