+++ internal/util/searchutil/scheduler/queues_test.go
@@ -1,4 +1,4 @@
-package tasks
+package scheduler
 
 import (
 	"fmt"
