+++ internal/util/searchutil/scheduler/tasks.go
@@ -1,4 +1,4 @@
-package tasks
+package scheduler
 
 import "github.com/milvus-io/milvus/internal/proto/internalpb"
