+++ internal/util/searchutil/scheduler/user_task_polling_policy.go
@@ -1,4 +1,4 @@
-package tasks
+package scheduler
 
 import (
 	"fmt"
